Refactor account registration and update database setup
Refactored the account registration process by removing the Register.kt and replacing it with AccountRouting.kt and RegisterRequest.kt. This change gives a clearer separation of responsibilities in the code. Adjusted the database setup methods in UsersSchema.kt, BearerSchema.kt and MagisterData.kt to create only missing tables and columns which offers better database management.